requirements.txt
discord.py==2.5.2
openai==1.99.1
